requirements.txt
antlr4-python3-runtime==4.13.2
antlr4-tools==0.2.2
install-jdk==1.1.0
pip==25.2
PyYAML==6.0.3